Feat: 하나의 스택에 단일 웹뷰만 사용하도록 변경
