+++ app/index.tsx
@@ -2,6 +2,7 @@ import { WEBVIEW_URL } from "@/constants/WebView";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { useRouter } from "expo-router";
 import { useEffect } from "react";
+import { Button } from "react-native";
 import { WebView, WebViewMessageEvent } from "react-native-webview";
 
 const SignInScreen = () => {
@@ -12,7 +13,7 @@ const SignInScreen = () => {
       const isLoggedIn = await AsyncStorage.getItem("isLoggedIn");
 
       if (isLoggedIn) {
-        router.replace("/(tabs)");
+        router.replace("/main");
       }
     };
 
@@ -23,16 +24,25 @@ const SignInScreen = () => {
     const message = JSON.parse(event.nativeEvent.data);
   };
   return (
-    <WebView
-      source={{ uri: `${WEBVIEW_URL}/signIn` }}
-      onMessage={handleMessage}
-      onNavigationStateChange={async (e) => {
-        if (e.url.includes("/signIn/redirect")) {
+    <>
+      <WebView
+        source={{ uri: `${WEBVIEW_URL}/signIn` }}
+        onMessage={handleMessage}
+        onNavigationStateChange={async (e) => {
+          if (e.url.includes("/signIn/redirect")) {
+            await AsyncStorage.setItem("isLoggedIn", "true");
+            router.replace("/main");
+          }
+        }}
+      />
+      <Button
+        onPress={async () => {
           await AsyncStorage.setItem("isLoggedIn", "true");
-          router.replace("/(tabs)");
-        }
-      }}
-    />
+        }}
+        title="앱진입"
+        color="#841584"
+      />
+    </>
   );
 };
