+++ app/main.tsx
@@ -1,17 +1,30 @@
 import { WebView, WebViewMessageEvent } from "react-native-webview";
+import * as Location from "expo-location";
+import { Alert } from "react-native";
 import { useRef } from "react";
 import useLocationPermission from "@/hooks/useLocationPermission";
+import { useRouter } from "expo-router";
 import { WEBVIEW_URL } from "@/constants/WebView";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { useNavigation } from "expo-router";
-import { CommonActions } from "@react-navigation/native";
 
-export default function MyScreen() {
+export default function HomeScreen() {
   const webViewRef = useRef<WebView>(null);
-  const navigation = useNavigation();
 
   const { checkLocationService, checkPermissions, requestPermissions } =
     useLocationPermission();
+
+  const router = useRouter();
+
+  const getLocation = async () => {
+    try {
+      const {
+        coords: { latitude, longitude },
+      } = await Location.getCurrentPositionAsync();
+      webViewRef.current?.postMessage(JSON.stringify({ latitude, longitude }));
+    } catch {
+      Alert.alert("위치 정보를 받아오는데 오류가 발생했습니다.");
+    }
+  };
 
   const handleMessage = async (event: WebViewMessageEvent) => {
     const message = JSON.parse(event.nativeEvent.data);
@@ -33,6 +46,12 @@ export default function MyScreen() {
           );
           return;
         }
+
+        getLocation();
+        break;
+      }
+      case "STACK_TRACKING": {
+        router.push("/tracking");
         break;
       }
       case "GPS_PERMISSION_STATE": {
@@ -46,23 +65,16 @@ export default function MyScreen() {
       case "LOG_OUT": {
         await AsyncStorage.removeItem("isLoggedIn");
 
-        navigation.dispatch(
-          CommonActions.reset({
-            index: 0,
-            routes: [{ name: "/" }], // 초기화 후 이동할 경로
-          })
-        );
+        router.replace("/");
       }
     }
   };
 
   return (
-    <>
-      <WebView
-        ref={webViewRef}
-        source={{ uri: `${WEBVIEW_URL}/mypage` }}
-        onMessage={handleMessage}
-      />
-    </>
+    <WebView
+      ref={webViewRef}
+      source={{ uri: `${WEBVIEW_URL}` }}
+      onMessage={handleMessage}
+    />
   );
 }
